Stripe error tests
Got rid of yellow border around alert. The text shows yellow if there's
an error; otherwise it's invisible. Tested error paths and redirected
to stay on order page and display more descriptive error message. There
are two kinds of errors: ones JS catches, which go into the alert box,
and ones the controller catches, which go into the flash messages. In
both cases it stays on the order page and lets the user fix it.