Make strategy terms updateable; misc other fixes
When showing payments for a vendor, only show deals (not ads!). Put
promotion terms in the vendor promotion status emails. Don't just print
the urls of website/facebook on the order page; make real links. Pass
relative and product strategy parameters in so that they reflect the
saved values! Messy!