Web and facebook links now work
JS was blocking the click; have to watch that for anything in a tab
pane. Also, they might not include the http:// prefix, which can make
the link not work (it converts it to a relative address, which doesn't
work). So create display url/facebook methods and call them instead of
reading it directly. Some vendors might include the http, others might
not. Allow either way.