Ensure voucher consistency
Fix the bug I encountered when extending the Fitness Factory deadline.
If you extend a promotion with a Fixed strategy *past* the end date of
the strategy, it will work -- but when somebody orders, it will
generate vouchers that don't validate (i.e., they would be created as
expired). So you have an order with no associated vouchers, and a mess
of trouble. But it lets you do it, so you don't know there's a problem
until somebody orders! This prevents saving the promotion if it's going
to do that.