Remove category filter from machovy_feed
Internal Erotica feed is now broken, because all the deals are in an
exclusive category, and the internal feed tries to respect categories.
I think if they select the vendor specifically, it should show all the
deals regardless of category (like the external feeds, including
midnightguru, which will continue to show the adult content). This is
probably more consistent with the principle of least surprise.