Pathological email cases
VI Enterprises is a vendor with no associated user. But the user mailer
assumes a vendor has an associated email to send the emails to. So at
this point the emails won't go out if anybody buys Steeler tickets!
It's possible to associate an admin user with this kind of vendor; it
doesn't break anything, but doesn't feel right. What is cleaner and has
the same effect is to change the mailer to check for this case and send
emails to the sales admin. Vince has a user account, but I'm not sure
he reads that email. Safer to send it to the admins.