Add tasks to handle After Hours category
I want to be able to toggle this setting back and forth easily! The
mixin_after_hours task mixes strip clubs into the general deals (and
adds them to NightLife). They are in After Hours *and* NightLife; After
Hours is no longer exclusive. The category is also set inactive, which
means it won't be displayed in the category bar -- but it is still
there and able to be edited. Do not remove the After Hours category in
this state, or it won't be reversible! The reverse task removes all
other categories assigned to After Hours promotions, and makes it once
again Active and Exclusive. If you want to leave After Hours mixed in
but still display it in the category bar, simply set Active to true.
Added an active scope, but am not using it now; we don't want to
restrict editing to active, or we'll lose the After Hours identity if
they ever edit it (e.g., to add something else), and it won't be
reversible.